Added recurrenceInterval to AddHabitModal and Habit.ts
