+++ habit_tracker/src/types/Recurrence.ts
@@ -1,1 +1,5 @@
 export type Recurrence = 'daily' | 'weekly' | 'monthly';
+
+export const isRecurrence = (value: string): value is Recurrence => {
+    return value === 'daily' || value === 'weekly' || value === 'monthly';
+}
